requirements.txt
Flask
docker
psutil
python-dotenv
fabric
pyyaml
tabulate
cryptography
pytest
